+++ store/src/test/java/org/xowl/infra/store/storage/persistent/FileBackendPerfTest.java
@@ -73,7 +73,7 @@ public class FileBackendPerfTest {
                         try {
                             for (int i = 0; i != BATCH_COUNT; i++) {
                                 for (int j = 0; j != WRITE_COUNT; j++) {
-                                    long key = pf.allocateDirect(WRITE_SIZE);
+                                    int key = pf.allocateDirect(WRITE_SIZE);
                                     try (IOAccess access = pf.access(key, true)) {
                                         for (int k = 0; k != WRITE_SIZE >> 2; k++) {
                                             access.writeInt(j);
